Prepare Inventory Methods
Preparing all the inventory methods to be general for reagents and
glasswares.

+ New ItemToInventory class added. All objects that might go into the
inventory should inherit this one
(The class is actually empty. It was made just to make possible the
polymorphism with Glasswares and Reagents)

OBS: This is pretty imcomplete because I lost what I was doing before.
I'll but everything back here as soons as possible.